chaincfg: Move simnet code to testnet files.
This moves the various code that is specific to simnet to the
simnetparams.go file to keep the information housed together.  Along the
same lines, it moves the tests specific to the simnet params to
simnetparams_test.go.

It also unexports the block one ledger definition since it should only
be exposed via the params.
